need to fix packet loss in can transmission

+++ projects/bootloader/inc/boot_flash.h
@@ -2,7 +2,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#include "global.h"
+#include "bootloader_mcu.h"
 #include "stm32f10x.h"
 
 #define NUM_FLASH_PAGES 64
@@ -11,7 +11,7 @@
   (((uintptr_t)(addr) - (uintptr_t)APP_START_ADDRESS) / BOOTLOADER_PAGE_BYTES)
 
 #define BOOTLOADER_PAGE_TO_ADDR(page) \
-  ((uintptr_t)(page)*BOOTLOADER_PAGE_BYTES + (uintptr_t)APP_START_ADDRESS)
+  ((uintptr_t)(page) * (uintptr_t)BOOTLOADER_PAGE_BYTES + (uintptr_t)APP_START_ADDRESS)
 #define BOOTLOADER_WRITE_BYTES 4
 #define BOOTLOADER_PAGE_BYTES 0x400
